// src/types/ActionTypes.ts
export enum ActionTypes {
	SET_DICE = 'SET_DICE',
	SET_NOTICE = 'SET_NOTICE',
	TOGGLE_CURRENT_PLAYER = 'TOGGLE_CURRENT_PLAYER',
	MOVE_CHECKER = 'MOVE_CHECKER',
	UNDO_MOVE = 'UNDO_MOVE',
	REDO_MOVE = 'REDO_MOVE',
	RESTART_GAME = 'RESTART_GAME',
	SURRENDER_GAME = 'SURRENDER_GAME',
	SAVE_GAME = 'SAVE_GAME',
	LOAD_GAME = 'LOAD_GAME',
	SET_GAME_OVER = 'SET_GAME_OVER',
	SHOW_GAME_HISTORY = 'SHOW_GAME_HISTORY',
	SHOW_GAME_STATISTICS = 'SHOW_GAME_STATISTICS',
	SHOW_PLAYER_STATISTICS = 'SHOW_PLAYER_STATISTICS',
	SHOW_PIP_COUNT = 'SHOW_PIP_COUNT',
	UPDATE_PIP_COUNT = 'UPDATE_PIP_COUNT',
	OFFER_DOUBLING_CUBE = 'OFFER_DOUBLING_CUBE',
	ACCEPT_DOUBLE = 'ACCEPT_DOUBLE',
	DROP_DOUBLE = 'DROP_DOUBLE',
	ROLL_OPENING_DIE = 'ROLL_OPENING_DIE',
	SET_MATCH_TARGET = 'SET_MATCH_TARGET',
}